Feature complete
SyncParticleSignal now accepts additional builder methods (with/without)
to allow filtering of which particle components should be synchronized.
The old API remains the same (all components will synchronize by
default).

The nature of these changes were mainly inspired by Bevy's
DynamicSceneBuilder pattern, which leverages TypeId for component
filtering when creating a DyanmicScene
