+++ src/core/particle/sync.rs
@@ -7,14 +7,18 @@
 //! - `Changed<Particle>` fires (e.g., on spawn or when [`Particle::name`] is mutated).
 //!
 //! The [`ParticleSyncExt`] trait provides an interface for adding your own particle sync
-//! components.
+//! components. Each registered propagator is keyed by a [`TypeId`], enabling
+//! selective synchronization through [`PropagatorFilter`]. Use
+//! [`SyncParticleSignal::with`]/[`SyncParticleSignal::without`] (and the corresponding methods
+//! on [`SyncParticleTypeChildrenSignal`]) to control which components are synchronized.
 use super::LocateBy;
 use crate::core::{
     AttachedToParticleType, ChanceLifetime, ChunkDirtyState, ChunkIndex, GridPosition, Particle,
     ParticleMap, ParticleSystems, ParticleType, ParticleTypeRegistry, TimedLifetime,
 };
-use bevy::{ecs::system::SystemParam, prelude::*};
+use bevy::{ecs::system::SystemParam, platform::collections::HashSet, prelude::*};
 use serde::{Deserialize, Serialize};
+use std::any::TypeId;
 
 pub(super) struct SyncPlugin;
 
@@ -46,8 +50,55 @@ impl Plugin for SyncPlugin {
 
 type PropagateFn = Box<dyn Fn(Entity, Entity, &mut Commands) + Send + Sync>;
 
+/// Controls which propagators run during particle synchronization.
+///
+/// Constructed implicitly via [`SyncParticleSignal::with`], [`SyncParticleSignal::without`],
+/// and the corresponding methods on [`SyncParticleTypeChildrenSignal`].
+#[derive(Clone, Debug, Default, Eq, PartialEq, Hash)]
+pub enum PropagatorFilter {
+    /// Run all registered propagators (the default).
+    #[default]
+    All,
+    /// Run only the propagators whose keys are in this list.
+    Only(Vec<TypeId>),
+    /// Run all propagators except those whose keys are in this list.
+    Except(Vec<TypeId>),
+}
+
 #[derive(Resource, Default)]
-pub(crate) struct ParticlePropagators(Vec<PropagateFn>);
+pub(crate) struct ParticlePropagators(Vec<(TypeId, PropagateFn)>);
+
+impl ParticlePropagators {
+    fn run_filtered(
+        &self,
+        filter: &PropagatorFilter,
+        entity: Entity,
+        parent: Entity,
+        commands: &mut Commands,
+    ) {
+        match filter {
+            PropagatorFilter::All => {
+                for (_, f) in &self.0 {
+                    f(entity, parent, commands);
+                }
+            }
+            PropagatorFilter::Only(ids) => {
+                for (key, f) in &self.0 {
+                    if ids.contains(key) {
+                        f(entity, parent, commands);
+                    }
+                }
+            }
+            PropagatorFilter::Except(ids) => {
+                for (key, f) in &self.0 {
+                    if !ids.contains(key) {
+                        f(entity, parent, commands);
+                    }
+                }
+            }
+        }
+    }
+}
 
 /// Extension trait for registering custom particle component propagation.
 ///
@@ -61,16 +112,21 @@ pub(crate) struct ParticlePropagators(Vec<PropagateFn>);
 /// mutated). Each propagator receives the particle entity, its parent [`ParticleType`] entity,
 /// and [`Commands`] for deferred mutations.
 ///
+/// Every propagator is keyed by a [`TypeId`], which enables selective
+/// synchronization via [`PropagatorFilter`]. Signals can target specific propagators with
+/// [`SyncParticleSignal::with`]/[`SyncParticleSignal::without`].
+///
 /// There are two registration methods:
 ///
 /// - [`register_particle_sync_component`](`ParticleSyncExt::register_particle_sync_component`):
 ///   Clones a component from the parent [`ParticleType`] onto each child [`Particle`]. If the
-///   parent doesn't have the component, it is removed from the child. This covers the majority of
-///   use cases.
+///   parent doesn't have the component, it is removed from the child. The component type `C` is
+///   used as the propagator key. This covers the majority of use cases.
 ///
 /// - [`register_particle_propagator`](`ParticleSyncExt::register_particle_propagator`): Accepts
 ///   an arbitrary closure for cases where a simple clone isn't enough (e.g. conditional insertion,
-///   deriving values from the parent, etc.).
+///   deriving values from the parent, etc.). Requires an explicit marker type `K` as the
+///   propagator key.
 ///
 /// After registration, updating the component on a [`ParticleType`] entity and sending a
 /// [`SyncParticleTypeChildrenSignal`] will propagate the change to all living particles of that
@@ -78,6 +134,9 @@ pub(crate) struct ParticlePropagators(Vec<PropagateFn>);
 pub trait ParticleSyncExt {
     /// Registers a component type to be cloned from [`ParticleType`] to child [`Particle`]
     /// entities.
+    ///
+    /// The component type `C` is used as both the source to clone and the propagator key for
+    /// [`PropagatorFilter`]-based targeting.
     ///
     /// When a particle is spawned or synced, if its parent [`ParticleType`] has the component `C`,
     /// a clone is inserted into the particle. If the parent does not have the component, it is
@@ -111,6 +170,10 @@ pub trait ParticleSyncExt {
 
     /// Registers an arbitrary propagator that runs for each particle during synchronization.
     ///
+    /// The type parameter `K` is used as a key for the propagator, allowing it to be targeted
+    /// by [`PropagatorFilter`]. Typically `K` is the parent component type that drives the
+    /// propagation logic.
+    ///
     /// The closure receives `(particle_entity, parent_entity, &mut Commands)` where
     /// `particle_entity` is the [`Particle`], `parent_entity` is its [`ParticleType`], and
     /// [`Commands`] is used for deferred mutations.
@@ -135,7 +198,7 @@ pub trait ParticleSyncExt {
     /// fn main() {
     ///     App::new()
     ///         .add_plugins((DefaultPlugins, FallingSandPlugin::default()))
-    ///         .register_particle_propagator(|particle, parent, commands| {
+    ///         .register_particle_propagator::<MaxHealth>(|particle, parent, commands| {
     ///             commands.queue(move |world: &mut World| {
     ///                 let max_hp = world.get::<MaxHealth>(parent).map(|h| h.0);
     ///                 if let Some(max_hp) = max_hp {
@@ -148,14 +211,15 @@ pub trait ParticleSyncExt {
     ///         .run();
     /// }
     /// ```
-    fn register_particle_propagator<F>(&mut self, propagator: F) -> &mut Self
-    where
-        F: Fn(Entity, Entity, &mut Commands) + Send + Sync + 'static;
+    fn register_particle_propagator<K: 'static>(
+        &mut self,
+        propagator: impl Fn(Entity, Entity, &mut Commands) + Send + Sync + 'static,
+    ) -> &mut Self;
 }
 
 impl ParticleSyncExt for App {
     fn register_particle_sync_component<C: Component + Clone>(&mut self) -> &mut Self {
-        self.register_particle_propagator(
+        self.register_particle_propagator::<C>(
             |entity: Entity, parent: Entity, commands: &mut Commands| {
                 commands.queue(move |world: &mut World| {
                     let component = world.get::<C>(parent).cloned();
@@ -169,14 +233,14 @@ impl ParticleSyncExt for App {
         )
     }
 
-    fn register_particle_propagator<F>(&mut self, propagator: F) -> &mut Self
-    where
-        F: Fn(Entity, Entity, &mut Commands) + Send + Sync + 'static,
-    {
+    fn register_particle_propagator<K: 'static>(
+        &mut self,
+        propagator: impl Fn(Entity, Entity, &mut Commands) + Send + Sync + 'static,
+    ) -> &mut Self {
         self.world_mut()
             .resource_mut::<ParticlePropagators>()
             .0
-            .push(Box::new(propagator));
+            .push((TypeId::of::<K>(), Box::new(propagator)));
         self
     }
 }
@@ -210,10 +274,8 @@ struct SyncParticleParams<'w, 's> {
 
 #[allow(clippy::needless_pass_by_value)]
 fn msgr_sync_particle(mut params: SyncParticleParams) {
-    use bevy::platform::collections::HashSet;
-
     let mut seen = HashSet::new();
-    let mut targets: Vec<(Entity, Entity, IVec2)> = Vec::new();
+    let mut targets: Vec<(Entity, Entity, IVec2, PropagatorFilter)> = Vec::new();
 
     for signal in params.msgr.read() {
         let entity = match &signal.locate_by {
@@ -224,7 +286,9 @@ fn msgr_sync_particle(mut params: SyncParticleParams) {
         if let Some(entity) = entity {
             if let Ok((attached, grid_pos)) = params.particle_query.get(entity) {
                 if seen.insert(entity) {
-                    targets.push((entity, attached.0, grid_pos.0));
+                    targets.push((entity, attached.0, grid_pos.0, signal.filter.clone()));
+                } else if let Some(target) = targets.iter_mut().find(|(e, _, _, _)| *e == entity) {
+                    target.3 = PropagatorFilter::All;
                 }
             }
         }
@@ -232,7 +296,9 @@ fn msgr_sync_particle(mut params: SyncParticleParams) {
 
     for (entity, attached, grid_pos) in &params.changed_query {
         if seen.insert(entity) {
-            targets.push((entity, attached.0, grid_pos.0));
+            targets.push((entity, attached.0, grid_pos.0, PropagatorFilter::All));
+        } else if let Some(target) = targets.iter_mut().find(|(e, _, _, _)| *e == entity) {
+            target.3 = PropagatorFilter::All;
         }
     }
 
@@ -240,13 +306,13 @@ fn msgr_sync_particle(mut params: SyncParticleParams) {
         return;
     }
 
-    for &(entity, parent, _) in &targets {
-        for propagate_fn in &params.propagators.0 {
-            propagate_fn(entity, parent, &mut params.commands);
-        }
+    for (entity, parent, _, ref filter) in &targets {
+        params
+            .propagators
+            .run_filtered(filter, *entity, *parent, &mut params.commands);
     }
 
-    for &(_, _, position) in &targets {
+    for &(_, _, position, _) in &targets {
         let chunk_coord = params.chunk_index.world_to_chunk_coord(position);
         if let Some(chunk_entity) = params.chunk_index.get(chunk_coord) {
             if let Ok(mut dirty_state) = params.chunk_dirty_query.get_mut(chunk_entity) {
@@ -282,9 +348,7 @@ fn on_sync_particle(
     let parent = attached.0;
     let position = grid_pos.0;
 
-    for propagate_fn in &propagators.0 {
-        propagate_fn(entity, parent, &mut commands);
-    }
+    propagators.run_filtered(&event.filter, entity, parent, &mut commands);
 
     let chunk_coord = chunk_index.world_to_chunk_coord(position);
     if let Some(chunk_entity) = chunk_index.get(chunk_coord) {
@@ -295,9 +359,16 @@ fn on_sync_particle(
 }
 
 /// Signal for synchronizing a [`Particle`] entity with its [`ParticleType`] parent.
+///
+/// By default, all registered propagators run. Use [`with`](Self::with) to restrict
+/// synchronization to specific components, or [`without`](Self::without) to exclude specific
+/// components.
 #[derive(Event, Message, Clone, Eq, PartialEq, Hash, Debug, Reflect, Serialize, Deserialize)]
 pub struct SyncParticleSignal {
     locate_by: LocateBy,
+    #[serde(skip)]
+    #[reflect(ignore)]
+    filter: PropagatorFilter,
 }
 
 impl SyncParticleSignal {
@@ -322,6 +393,7 @@ impl SyncParticleSignal {
     pub const fn from_entity(entity: Entity) -> Self {
         Self {
             locate_by: LocateBy::Entity(entity),
+            filter: PropagatorFilter::All,
         }
     }
 
@@ -341,15 +413,84 @@ impl SyncParticleSignal {
     pub const fn from_position(position: IVec2) -> Self {
         Self {
             locate_by: LocateBy::Position(position),
+            filter: PropagatorFilter::All,
         }
+    }
+
+    /// Narrow synchronization to include the propagator keyed by `C`.
+    ///
+    /// When called on a signal with no filter ([`PropagatorFilter::All`]), switches to
+    /// [`PropagatorFilter::Only`] containing `C`. When called on an existing
+    /// [`PropagatorFilter::Only`], adds `C`. When called on [`PropagatorFilter::Except`],
+    /// removes `C` from the exclusion list.
+    ///
+    /// Chain multiple calls to sync several components:
+    ///
+    /// ```ignore
+    /// SyncParticleSignal::from_entity(e).with::<Density>().with::<Speed>()
+    /// ```
+    #[must_use]
+    pub fn with<C: 'static>(mut self) -> Self {
+        let id = TypeId::of::<C>();
+        match &mut self.filter {
+            PropagatorFilter::All => {
+                self.filter = PropagatorFilter::Only(vec![id]);
+            }
+            PropagatorFilter::Only(ids) => {
+                if !ids.contains(&id) {
+                    ids.push(id);
+                }
+            }
+            PropagatorFilter::Except(ids) => {
+                ids.retain(|i| *i != id);
+            }
+        }
+        self
+    }
+
+    /// Exclude the propagator keyed by `C` from synchronization.
+    ///
+    /// When called on a signal with no filter ([`PropagatorFilter::All`]), switches to
+    /// [`PropagatorFilter::Except`] excluding `C`. When called on an existing
+    /// [`PropagatorFilter::Except`], adds `C` to the exclusion list. When called on
+    /// [`PropagatorFilter::Only`], removes `C` from the inclusion list.
+    ///
+    /// ```ignore
+    /// SyncParticleSignal::from_entity(e).without::<ColorProfile>()
+    /// ```
+    #[must_use]
+    pub fn without<C: 'static>(mut self) -> Self {
+        let id = TypeId::of::<C>();
+        match &mut self.filter {
+            PropagatorFilter::All => {
+                self.filter = PropagatorFilter::Except(vec![id]);
+            }
+            PropagatorFilter::Except(ids) => {
+                if !ids.contains(&id) {
+                    ids.push(id);
+                }
+            }
+            PropagatorFilter::Only(ids) => {
+                ids.retain(|i| *i != id);
+            }
+        }
+        self
     }
 }
 
 /// Signal for syncing all [`Particle`] entities of a matching
-/// [`ParticleType`]
+/// [`ParticleType`].
+///
+/// By default, all registered propagators run. Use [`with`](Self::with) to restrict
+/// synchronization to specific components, or [`without`](Self::without) to exclude specific
+/// components. The filter is forwarded to each individual [`SyncParticleSignal`] created for
+/// the children.
 #[derive(Event, Message, Clone, Eq, PartialEq, Hash, Debug, Reflect, Serialize, Deserialize)]
 pub struct SyncParticleTypeChildrenSignal {
     locate_by: LocateBy,
+    #[serde(skip)]
+    #[reflect(ignore)]
+    filter: PropagatorFilter,
 }
 
 impl SyncParticleTypeChildrenSignal {
@@ -369,6 +510,7 @@ impl SyncParticleTypeChildrenSignal {
     pub const fn from_name(name: String) -> Self {
         Self {
             locate_by: LocateBy::Name(name),
+            filter: PropagatorFilter::All,
         }
     }
 
@@ -393,7 +535,52 @@ impl SyncParticleTypeChildrenSignal {
     pub const fn from_parent_handle(entity: Entity) -> Self {
         Self {
             locate_by: LocateBy::Entity(entity),
+            filter: PropagatorFilter::All,
         }
+    }
+
+    /// Narrow synchronization to include the propagator keyed by `C`.
+    ///
+    /// See [`SyncParticleSignal::with`] for details on filter behavior.
+    #[must_use]
+    pub fn with<C: 'static>(mut self) -> Self {
+        let id = TypeId::of::<C>();
+        match &mut self.filter {
+            PropagatorFilter::All => {
+                self.filter = PropagatorFilter::Only(vec![id]);
+            }
+            PropagatorFilter::Only(ids) => {
+                if !ids.contains(&id) {
+                    ids.push(id);
+                }
+            }
+            PropagatorFilter::Except(ids) => {
+                ids.retain(|i| *i != id);
+            }
+        }
+        self
+    }
+
+    /// Exclude the propagator keyed by `C` from synchronization.
+    ///
+    /// See [`SyncParticleSignal::without`] for details on filter behavior.
+    #[must_use]
+    pub fn without<C: 'static>(mut self) -> Self {
+        let id = TypeId::of::<C>();
+        match &mut self.filter {
+            PropagatorFilter::All => {
+                self.filter = PropagatorFilter::Except(vec![id]);
+            }
+            PropagatorFilter::Except(ids) => {
+                if !ids.contains(&id) {
+                    ids.push(id);
+                }
+            }
+            PropagatorFilter::Only(ids) => {
+                ids.retain(|i| *i != id);
+            }
+        }
+        self
     }
 }
 
@@ -419,7 +606,9 @@ fn msgr_sync_particle_type_children(
                 .iter()
                 .filter(|(_, attached)| attached.0 == *parent)
             {
-                msgr_sync_particle.write(SyncParticleSignal::from_entity(entity));
+                let mut signal = SyncParticleSignal::from_entity(entity);
+                signal.filter = msg.filter.clone();
+                msgr_sync_particle.write(signal);
             }
         } else {
             warn!("No particle type found in particle type registry!");
@@ -449,7 +638,9 @@ fn on_sync_particle_type_children(
             .iter()
             .filter(|(_, attached)| attached.0 == *parent)
         {
-            msgr_sync_particle.write(SyncParticleSignal::from_entity(entity));
+            let mut signal = SyncParticleSignal::from_entity(entity);
+            signal.filter = event.filter.clone();
+            msgr_sync_particle.write(signal);
         }
     } else {
         warn!("No particle type found in particle type registry!");
@@ -533,5 +724,637 @@ fn sync_particle_parent(
                 .name
                 .clone_from(&parent_type.name);
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{
+        core::{
+            AttachedToParticleType, Particle, ParticleType, ParticleTypeRegistry,
+            SpawnParticleSignal,
+        },
+        FallingSandMinimalPlugin,
+    };
+
+    #[derive(Component, Clone, Debug, PartialEq)]
+    struct Marker(u32);
+
+    #[derive(Component, Clone, Debug, PartialEq)]
+    struct Marker2(u32);
+
+    fn create_test_app() -> App {
+        let mut app = App::new();
+        app.add_plugins(MinimalPlugins);
+        app.add_plugins(FallingSandMinimalPlugin::default());
+        app
+    }
+
+    // ---- sync_particle_type_registry ----
+
+    #[test]
+    fn sync_particle_type_registry_on_name_mutation() {
+        let mut app = create_test_app();
+
+        let entity = app.world_mut().spawn(ParticleType::new("sand")).id();
+        app.update();
+
+        let registry = app.world().resource::<ParticleTypeRegistry>();
+        assert_eq!(registry.get("sand"), Some(&entity));
+
+        app.world_mut()
+            .entity_mut(entity)
+            .get_mut::<ParticleType>()
+            .unwrap()
+            .name = "water".into();
+        app.update();
+
+        let registry = app.world().resource::<ParticleTypeRegistry>();
+        assert_eq!(registry.get("water"), Some(&entity));
+        assert_eq!(registry.get("sand"), None);
+    }
+
+    // ---- sync_particle_parent ----
+
+    #[test]
+    fn sync_particle_parent_updates_attached_on_particle_mutation() {
+        let mut app = create_test_app();
+
+        let sand_pt = app.world_mut().spawn(ParticleType::new("sand")).id();
+        let water_pt = app.world_mut().spawn(ParticleType::new("water")).id();
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        let attached = app
+            .world()
+            .entity(particle_entity)
+            .get::<AttachedToParticleType>()
+            .unwrap();
+        assert_eq!(attached.0, sand_pt);
+
+        app.world_mut()
+            .entity_mut(particle_entity)
+            .get_mut::<Particle>()
+            .unwrap()
+            .name = "water".into();
+        app.update();
+        app.update();
+
+        let attached = app
+            .world()
+            .entity(particle_entity)
+            .get::<AttachedToParticleType>()
+            .unwrap();
+        assert_eq!(attached.0, water_pt);
+    }
+
+    #[test]
+    fn sync_particle_parent_reverts_unregistered_name() {
+        let mut app = create_test_app();
+
+        let sand_pt = app.world_mut().spawn(ParticleType::new("sand")).id();
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        app.world_mut()
+            .entity_mut(particle_entity)
+            .get_mut::<Particle>()
+            .unwrap()
+            .name = "ghost".into();
+        app.update();
+
+        let particle = app
+            .world()
+            .entity(particle_entity)
+            .get::<Particle>()
+            .unwrap();
+        assert_eq!(
+            particle.name, "sand",
+            "Name should revert to parent ParticleType"
+        );
+
+        let attached = app
+            .world()
+            .entity(particle_entity)
+            .get::<AttachedToParticleType>()
+            .unwrap();
+        assert_eq!(attached.0, sand_pt);
+    }
+
+    // ---- sync_registered_components ----
+
+    #[test]
+    fn sync_registered_components_propagates_on_spawn() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+
+        app.world_mut()
+            .spawn((ParticleType::new("sand"), Marker(42)));
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        let marker = app
+            .world()
+            .entity(particle_entity)
+            .get::<Marker>()
+            .expect("Marker should have been propagated from ParticleType");
+        assert_eq!(marker, &Marker(42));
+    }
+
+    #[test]
+    fn sync_registered_components_removes_absent_component() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+
+        app.world_mut()
+            .spawn((ParticleType::new("sand"), Marker(42)));
+        app.world_mut().spawn(ParticleType::new("water"));
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        assert!(app
+            .world()
+            .entity(particle_entity)
+            .get::<Marker>()
+            .is_some());
+
+        app.world_mut()
+            .entity_mut(particle_entity)
+            .get_mut::<Particle>()
+            .unwrap()
+            .name = "water".into();
+        app.update();
+        app.update();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            None,
+            "Marker should be removed when switching to a ParticleType that lacks it"
+        );
+    }
+
+    #[test]
+    fn sync_registered_components_updates_on_particle_type_change() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+
+        app.world_mut()
+            .spawn((ParticleType::new("sand"), Marker(1)));
+        app.world_mut()
+            .spawn((ParticleType::new("water"), Marker(2)));
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            Some(&Marker(1))
+        );
+
+        app.world_mut()
+            .entity_mut(particle_entity)
+            .get_mut::<Particle>()
+            .unwrap()
+            .name = "water".into();
+        app.update();
+        app.update();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            Some(&Marker(2)),
+            "Marker should reflect the new ParticleType's value"
+        );
+    }
+
+    // ---- sync_particle_signal ----
+
+    #[test]
+    fn msgr_sync_particle_triggers_change_detection_by_entity() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+
+        let pt_entity = app
+            .world_mut()
+            .spawn((ParticleType::new("sand"), Marker(1)))
+            .id();
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            Some(&Marker(1))
+        );
+
+        app.world_mut().entity_mut(pt_entity).insert(Marker(99));
+
+        app.update();
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            Some(&Marker(1))
+        );
+
+        app.world_mut()
+            .write_message(SyncParticleSignal::from_entity(particle_entity));
+        app.update();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            Some(&Marker(99)),
+            "SyncParticleSignal should trigger re-propagation of registered components"
+        );
+    }
+
+    #[test]
+    fn msgr_sync_particle_triggers_change_detection_by_position() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+
+        let pt_entity = app
+            .world_mut()
+            .spawn((ParticleType::new("sand"), Marker(1)))
+            .id();
+        app.update();
+
+        let position = IVec2::new(5, 5);
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), position));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        app.world_mut().entity_mut(pt_entity).insert(Marker(50));
+        app.update();
+
+        app.world_mut()
+            .write_message(SyncParticleSignal::from_position(position));
+        app.update();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            Some(&Marker(50)),
+        );
+    }
+
+    // ---- msgr_sync_particle_type_children ----
+
+    #[test]
+    fn msgr_sync_particle_type_children_by_name() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+
+        let pt_entity = app
+            .world_mut()
+            .spawn((ParticleType::new("sand"), Marker(1)))
+            .id();
+        app.update();
+
+        for i in 0..5 {
+            app.world_mut().write_message(SpawnParticleSignal::new(
+                Particle::new("sand"),
+                IVec2::new(i, 0),
+            ));
+        }
+        app.update();
+
+        let particle_entities: Vec<Entity> = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .collect();
+        assert_eq!(particle_entities.len(), 5);
+
+        app.world_mut().entity_mut(pt_entity).insert(Marker(77));
+        app.update();
+
+        for &e in &particle_entities {
+            assert_eq!(app.world().entity(e).get::<Marker>(), Some(&Marker(1)));
+        }
+
+        app.world_mut()
+            .write_message(SyncParticleTypeChildrenSignal::from_name("sand".into()));
+        app.update();
+        app.update();
+
+        for &e in &particle_entities {
+            assert_eq!(
+                app.world().entity(e).get::<Marker>(),
+                Some(&Marker(77)),
+                "All children should be synced after SyncParticleTypeChildrenSignal"
+            );
+        }
+    }
+
+    #[test]
+    fn msgr_sync_particle_type_children_by_entity() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+
+        let pt_entity = app
+            .world_mut()
+            .spawn((ParticleType::new("sand"), Marker(1)))
+            .id();
+        app.update();
+
+        for i in 0..3 {
+            app.world_mut().write_message(SpawnParticleSignal::new(
+                Particle::new("sand"),
+                IVec2::new(i, 0),
+            ));
+        }
+        app.update();
+
+        let particle_entities: Vec<Entity> = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .collect();
+
+        app.world_mut().entity_mut(pt_entity).insert(Marker(33));
+        app.update();
+
+        app.world_mut()
+            .write_message(SyncParticleTypeChildrenSignal::from_parent_handle(
+                pt_entity,
+            ));
+        app.update();
+        app.update();
+
+        for &e in &particle_entities {
+            assert_eq!(app.world().entity(e).get::<Marker>(), Some(&Marker(33)));
+        }
+    }
+
+    #[test]
+    fn msgr_sync_particle_type_children_only_affects_matching_type() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+
+        let sand_pt = app
+            .world_mut()
+            .spawn((ParticleType::new("sand"), Marker(1)))
+            .id();
+        app.world_mut()
+            .spawn((ParticleType::new("water"), Marker(2)));
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.world_mut().write_message(SpawnParticleSignal::new(
+            Particle::new("water"),
+            IVec2::new(1, 0),
+        ));
+        app.update();
+
+        app.world_mut().entity_mut(sand_pt).insert(Marker(99));
+        app.update();
+
+        app.world_mut()
+            .write_message(SyncParticleTypeChildrenSignal::from_name("sand".into()));
+        app.update();
+        app.update();
+
+        let particles: Vec<(Particle, Marker)> = app
+            .world_mut()
+            .query::<(&Particle, &Marker)>()
+            .iter(app.world())
+            .map(|(p, m)| (p.clone(), m.clone()))
+            .collect();
+
+        let sand_marker = particles
+            .iter()
+            .find(|(p, _)| p.name == "sand")
+            .map(|(_, m)| m);
+        let water_marker = particles
+            .iter()
+            .find(|(p, _)| p.name == "water")
+            .map(|(_, m)| m);
+
+        assert_eq!(sand_marker, Some(&Marker(99)));
+        assert_eq!(
+            water_marker,
+            Some(&Marker(2)),
+            "Water particle should be unaffected"
+        );
+    }
+
+    // ---- propagator_filter ----
+
+    #[test]
+    fn sync_with_filter_only_syncs_specified_component() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+        app.register_particle_sync_component::<Marker2>();
+
+        app.world_mut()
+            .spawn((ParticleType::new("sand"), Marker(1), Marker2(2)));
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            Some(&Marker(1))
+        );
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker2>(),
+            Some(&Marker2(2))
+        );
+
+        app.world_mut()
+            .entity_mut(particle_entity)
+            .remove::<Marker>();
+        app.world_mut()
+            .entity_mut(particle_entity)
+            .remove::<Marker2>();
+
+        app.world_mut()
+            .write_message(SyncParticleSignal::from_entity(particle_entity).with::<Marker>());
+        app.update();
+        app.update();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            Some(&Marker(1)),
+            "Marker should be synced by the with::<Marker>() filter"
+        );
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker2>(),
+            None,
+            "Marker2 should NOT be synced because it was not included in the filter"
+        );
+    }
+
+    #[test]
+    fn sync_without_filter_excludes_specified_component() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+        app.register_particle_sync_component::<Marker2>();
+
+        app.world_mut()
+            .spawn((ParticleType::new("sand"), Marker(1), Marker2(2)));
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.update();
+
+        let particle_entity = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .next()
+            .unwrap();
+
+        app.world_mut()
+            .entity_mut(particle_entity)
+            .remove::<Marker>();
+        app.world_mut()
+            .entity_mut(particle_entity)
+            .remove::<Marker2>();
+
+        app.world_mut()
+            .write_message(SyncParticleSignal::from_entity(particle_entity).without::<Marker>());
+        app.update();
+        app.update();
+
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker>(),
+            None,
+            "Marker should NOT be synced because it was excluded by the filter"
+        );
+        assert_eq!(
+            app.world().entity(particle_entity).get::<Marker2>(),
+            Some(&Marker2(2)),
+            "Marker2 should be synced because it was not excluded"
+        );
+    }
+
+    #[test]
+    fn sync_children_signal_propagates_filter() {
+        let mut app = create_test_app();
+        app.register_particle_sync_component::<Marker>();
+        app.register_particle_sync_component::<Marker2>();
+
+        app.world_mut()
+            .spawn((ParticleType::new("sand"), Marker(10), Marker2(20)));
+        app.update();
+
+        app.world_mut()
+            .write_message(SpawnParticleSignal::new(Particle::new("sand"), IVec2::ZERO));
+        app.world_mut().write_message(SpawnParticleSignal::new(
+            Particle::new("sand"),
+            IVec2::new(1, 0),
+        ));
+        app.update();
+
+        let particle_entities: Vec<Entity> = app
+            .world_mut()
+            .query_filtered::<Entity, With<Particle>>()
+            .iter(app.world())
+            .collect();
+        assert_eq!(particle_entities.len(), 2);
+
+        for &entity in &particle_entities {
+            app.world_mut().entity_mut(entity).remove::<Marker>();
+            app.world_mut().entity_mut(entity).remove::<Marker2>();
+        }
+
+        app.world_mut().write_message(
+            SyncParticleTypeChildrenSignal::from_name("sand".into()).with::<Marker>(),
+        );
+        app.update();
+        app.update();
+
+        for &entity in &particle_entities {
+            assert_eq!(
+                app.world().entity(entity).get::<Marker>(),
+                Some(&Marker(10)),
+                "Marker should be synced via children signal filter"
+            );
+            assert_eq!(
+                app.world().entity(entity).get::<Marker2>(),
+                None,
+                "Marker2 should NOT be synced — excluded by children signal filter"
+            );
+        }
+    }
+
+    #[test]
+    fn propagator_filter_default_is_all() {
+        assert_eq!(PropagatorFilter::default(), PropagatorFilter::All);
     }
 }
